refactor: modified CJM to ESM
